pyproject.toml
[project]
name = "holo-chan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "litellm>=1.81.1",
    "numpy>=2.4.2",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "pyzmq>=27.1.0",
    "simpleaudio>=1.0.4",
    "sounddevice>=0.5.5",
]

[project.optional-dependencies]
discord = [
    "discord.py[voice]>=2.4.0",
    "discord-ext-voice-recv>=0.5.2a179",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
